2.0 rc

+++ wireless_connect.sh
@@ -34,35 +34,21 @@ if [ ! -z "$ANDROID_DEVICES" ]; then
 	#sleep 5
 	#adb devices
 
+        # install appium apk
+        if [ -f /usr/lib/node_modules/appium/node_modules/appium-uiautomator2-driver/node_modules/io.appium.settings/apks/settings_apk-debug.apk ]; then
+            adb install /usr/lib/node_modules/appium/node_modules/appium-uiautomator2-driver/node_modules/io.appium.settings/apks/settings_apk-debug.apk
+        fi
+
 	# download and install chrome apk from https://www.apkmirror.com/apk/google-inc/chrome/chrome-99-0-4844-73-release/
-	# version: x86 + x86_64 
+	# version: x86 + x86_64
 	# url: https://www.apkmirror.com/apk/google-inc/chrome/chrome-99-0-4844-73-release/google-chrome-fast-secure-99-0-4844-73-10-android-apk-download/
-
 	# /tmp/zebrunner/chrome/latest.apk is default shared location for chrome browser apk
-	adb install /tmp/zebrunner/chrome/latest.apk
-
-	# install appium apk
-	adb install /usr/lib/node_modules/appium/node_modules/io.appium.settings/apks/settings_apk-debug.apk
-
-        # switch to root account for running adb
-        adb root
-
-        sleep 5
-        ret=1
-        redroidDevice="device:5555"
-        while [[ $ret -eq 1 ]]; do
-            echo "Connecting as root to: ${redroidDevice}"
-            adb connect ${redroidDevice}
-            adb devices | grep ${redroidDevice} | grep "device"
-            ret=$?
-            if [[ $ret -eq 1 ]]; then
-                sleep ${REMOTE_ADB_POLLING_SEC}
-            fi
-        done
-        echo "Connected as root to: ${redroidDevice}."
-
-        adb devices
+        if [ -f /tmp/zebrunner/chrome/latest.apk ]; then
+	    adb install /tmp/zebrunner/chrome/latest.apk
+        fi
     fi
+
+    adb devices
 fi
